Added new functionality to be able to render links in error messages from the back end.
Error messages are now pre-rendered with meta-data using hogan before being sent to a view.

Updated a typo in the target which was 'targer' to make the window open in the same tab/window used the common window target of "DRHELPPAGE" i.e. target="DRHELPPAGE" this will open links that use that target in the same tab/window.