Fix for issue #50
If the file element visibility is set to hidden or display: none the file dialogue is not shown. So the file element is now initially loaded as a hidden element then dynamically set to a type of "file" when the page is ready. This stops the odd behavior with the flashing native file control/widget when the page loads and still allows users to pick a file.